# core-banking-ledger-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-banking-ledger
  description: Banking Ledger Core Application
  contact:
    name: Firefly Software Solutions Inc
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.catalis.vc/ledger
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/accounts/{accountId}/statements/{statementId}:
    get:
      tags:
        - Account Statements
      summary: Get Account Statement
      description: Retrieve a specific statement by ID.
      operationId: getStatement
      parameters:
        - name: accountId
          in: path
          description: Account ID
          required: true
          schema:
            type: string
            format: uuid
        - name: statementId
          in: path
          description: Statement ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Statement retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementMetadataDTO'
  /api/v1/accounts/{accountId}/statements/date-range:
    get:
      tags:
        - Account Statements
      summary: List Account Statements by Date Range
      description: Retrieve a paginated list of statements for a specific account within a date range.
      operationId: listStatementsByDateRange
      parameters:
        - name: accountId
          in: path
          description: Account ID
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          description: Start date (ISO format)
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date (ISO format)
          required: true
          schema:
            type: string
            format: date
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: Account statements retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
  /api/v1/accounts/{accountId}/statements:
    get:
      tags:
        - Account Statements
      summary: List Account Statements
      description: Retrieve a paginated list of statements for a specific account.
      operationId: listStatements
      parameters:
        - name: accountId
          in: path
          description: Account ID
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: Account statements retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
...
  /api/v1/account-spaces/{accountSpaceId}/statements/{statementId}:
    get:
      tags:
        - Account Space Statements
      summary: Get Account Space Statement
      description: Retrieve a specific statement by ID.
      operationId: getStatement_1
      parameters:
        - name: accountSpaceId
          in: path
          description: Account Space ID
          required: true
          schema:
            type: string
            format: uuid
        - name: statementId
          in: path
          description: Statement ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Statement retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementMetadataDTO'
  /api/v1/account-spaces/{accountSpaceId}/statements/date-range:
    get:
      tags:
        - Account Space Statements
      summary: List Account Space Statements by Date Range
      description: Retrieve a paginated list of statements for a specific account space within a date range.
      operationId: listStatementsByDateRange_1
      parameters:
        - name: accountSpaceId
          in: path
          description: Account Space ID
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          description: Start date (ISO format)
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date (ISO format)
          required: true
          schema:
            type: string
            format: date
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: Account space statements retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
  /api/v1/account-spaces/{accountSpaceId}/statements:
    get:
      tags:
        - Account Space Statements
      summary: List Account Space Statements
      description: Retrieve a paginated list of statements for a specific account space.
      operationId: listStatements_1
      parameters:
        - name: accountSpaceId
          in: path
          description: Account Space ID
          required: true
          schema:
            type: string
            format: uuid
        - name: pageNumber
          in: query
          description: The zero-based page number to retrieve.
          required: false
          schema:
            type: string
            description: The zero-based page number to retrieve.
            example: 0
            default: '0'
          example: 0
        - name: pageSize
          in: query
          description: The number of items per page.
          required: false
          schema:
            type: string
            description: The number of items per page.
            example: 10
            default: '10'
          example: 10
        - name: sortBy
          in: query
          description: The field to sort the results by.
          required: false
          schema:
            type: string
            description: The field to sort the results by.
            example: name
          example: name
        - name: sortDirection
          in: query
          description: The direction of sorting, either ASC or DESC.
          required: false
          schema:
            type: string
            description: The direction of sorting, either ASC or DESC.
            example: DESC
            default: DESC
          example: DESC
      responses:
        '200':
          description: Account space statements retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
components:
  schemas:
...
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    StatementMetadataDTO:
      required:
        - endDate
        - generationDate
        - includedDetails
        - includedPending
        - periodType
        - startDate
        - transactionCount
      type: object
      properties:
        dateCreated:
          type: string
          format: date-time
          readOnly: true
        dateUpdated:
          type: string
          format: date-time
          readOnly: true
        statementId:
          type: string
          format: uuid
          readOnly: true
        accountId:
          type: string
          format: uuid
        accountSpaceId:
          type: string
          format: uuid
        periodType:
          type: string
          enum:
            - MONTHLY
            - QUARTERLY
            - YEARLY
            - CUSTOM
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        generationDate:
          type: string
          format: date-time
        transactionCount:
          minimum: 0
          type: integer
          format: int32
        includedPending:
          type: boolean
        includedDetails:
          type: boolean